Functionnal ft_dtoa.c. FINALLY ! Still need's to be cleaned though.

+++ srcs/utils/ft_power.c
@@ -6,12 +6,12 @@
 /*   By: nplieger <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/17 09:51:37 by nplieger          #+#    #+#             */
-/*   Updated: 2022/11/17 12:40:13 by nplieger         ###   ########.fr       */
+/*   Updated: 2022/11/18 10:22:27 by nplieger         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "ft_printf.h"
 
-int	ft_power(int nb, int power)
+long long int	ft_power(long long int nb, int power)
 {
 	if (power < 0)
 		return (0);
